Gha kubernetes tests

+++ tests/playwright/src/specs/deploy-to-kubernetes.spec.ts
@@ -42,7 +42,7 @@ const CONTAINER_START_PARAMS: ContainerInteractiveParams = {
   attachTerminal: false,
 };
 
-const skipKindInstallation = process.env.SKIP_KIND_INSTALL ? process.env.SKIP_KIND_INSTALL : false;
+const skipKindInstallation = process.env.SKIP_KIND_INSTALL === 'true';
 
 test.beforeAll(async ({ runner, welcomePage, page, navigationBar }) => {
   test.setTimeout(350_000);
@@ -56,25 +56,27 @@ test.beforeAll(async ({ runner, welcomePage, page, navigationBar }) => {
 
     await ensureCliInstalled(page, 'Kind');
   }
-  await createKindCluster(page, CLUSTER_NAME, true, CLUSTER_CREATION_TIMEOUT);
+
+  if (process.env.GITHUB_ACTIONS && process.env.RUNNER_OS === 'Linux') {
+    await createKindCluster(page, CLUSTER_NAME, false, CLUSTER_CREATION_TIMEOUT, { useIngressController: false });
+  } else {
+    await createKindCluster(page, CLUSTER_NAME, true, CLUSTER_CREATION_TIMEOUT);
+  }
 });
 
 test.afterAll(async ({ runner, page }) => {
   test.setTimeout(90000);
   try {
-    await deleteContainer(page, CONTAINER_NAME);
-    await deleteImage(page, IMAGE_TO_PULL);
-    await deleteKindCluster(page, KIND_CONTAINER_NAME, CLUSTER_NAME);
+    if (!(process.env.GITHUB_ACTIONS && process.env.RUNNER_OS === 'Linux')) {
+      await deleteContainer(page, CONTAINER_NAME);
+      await deleteImage(page, IMAGE_TO_PULL);
+      await deleteKindCluster(page, KIND_CONTAINER_NAME, CLUSTER_NAME);
+    }
   } finally {
     await runner.close();
     console.log('Runner closed');
   }
 });
-
-test.skip(
-  !!process.env.GITHUB_ACTIONS && process.env.RUNNER_OS === 'Linux',
-  'Tests suite should not run on Linux platform',
-);
 
 test.describe.serial('Deploy a container to the Kind cluster', { tag: '@k8s_e2e' }, () => {
   test('Pull an image and start a container', async ({ navigationBar }) => {
