+++ tests/playwright/src/specs/kubernetes-edit-yaml.spec.ts
@@ -62,22 +62,24 @@ test.beforeAll(async ({ runner, welcomePage, page, navigationBar }) => {
 
     await ensureCliInstalled(page, 'Kind');
   }
-  await createKindCluster(page, CLUSTER_NAME, true, CLUSTER_CREATION_TIMEOUT);
+
+  if (process.env.GITHUB_ACTIONS && process.env.RUNNER_OS === 'Linux') {
+    await createKindCluster(page, CLUSTER_NAME, false, CLUSTER_CREATION_TIMEOUT, { useIngressController: false });
+  } else {
+    await createKindCluster(page, CLUSTER_NAME, true, CLUSTER_CREATION_TIMEOUT);
+  }
 });
 
 test.afterAll(async ({ runner, page }) => {
   test.setTimeout(90000);
   try {
-    await deleteKindCluster(page, KIND_NODE, CLUSTER_NAME);
+    if (!(process.env.GITHUB_ACTIONS && process.env.RUNNER_OS === 'Linux')) {
+      await deleteKindCluster(page, KIND_NODE, CLUSTER_NAME);
+    }
   } finally {
     await runner.close();
   }
 });
-
-test.skip(
-  !!process.env.GITHUB_ACTIONS && process.env.RUNNER_OS === 'Linux',
-  'Tests suite should not run on Linux platform',
-);
 
 test.describe.serial('Kubernetes Edit YAML Feature E2E Test', { tag: '@k8s_e2e' }, () => {
   test('Create a Kubernetes deployment resource', async ({ navigationBar }) => {
